refactor: split PlayerScreen into smaller components
